fix: clear database on different user login

+++ app/src/main/java/org/fossasia/openevent/app/data/LoginModel.java
@@ -1,12 +1,14 @@
 package org.fossasia.openevent.app.data;
 
+import android.support.annotation.VisibleForTesting;
+
 import org.fossasia.openevent.app.data.contract.ILoginModel;
 import org.fossasia.openevent.app.data.contract.IUtilModel;
 import org.fossasia.openevent.app.data.db.contract.IDatabaseRepository;
 import org.fossasia.openevent.app.data.models.Login;
-import org.fossasia.openevent.app.data.models.LoginResponse;
 import org.fossasia.openevent.app.data.models.User;
 import org.fossasia.openevent.app.data.network.EventService;
+import org.fossasia.openevent.app.data.repository.contract.IEventRepository;
 import org.fossasia.openevent.app.main.MainActivity;
 import org.fossasia.openevent.app.utils.Constants;
 import org.fossasia.openevent.app.utils.JWTUtils;
@@ -14,47 +16,66 @@ import org.fossasia.openevent.app.utils.JWTUtils;
 import javax.inject.Inject;
 
 import io.reactivex.Completable;
-import io.reactivex.Observable;
+import io.reactivex.Single;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.schedulers.Schedulers;
 
 public class LoginModel implements ILoginModel {
 
-    private IUtilModel utilModel;
-    private EventService eventService;
-    private IDatabaseRepository databaseRepository;
+    private static final User EMPTY = new User();
+
+    private final IUtilModel utilModel;
+    private final EventService eventService;
+    private final IEventRepository eventRepository;
+    private final IDatabaseRepository databaseRepository;
 
     @Inject
-    public LoginModel(IUtilModel utilModel, EventService eventService, IDatabaseRepository databaseRepository) {
+    public LoginModel(IUtilModel utilModel, EventService eventService, IEventRepository eventRepository, IDatabaseRepository databaseRepository) {
         this.utilModel = utilModel;
         this.eventService = eventService;
+        this.eventRepository = eventRepository;
         this.databaseRepository = databaseRepository;
     }
 
     @Override
-    public Observable<LoginResponse> login(String username, String password) {
-        if(isLoggedIn()) {
-            return Observable.just(new LoginResponse(utilModel.getToken()));
-        }
+    public Completable login(String username, String password) {
+        if(isLoggedIn())
+            return Completable.complete();
 
-        if(!utilModel.isConnected()) {
-            return Observable.error(new RuntimeException(Constants.NO_NETWORK));
-        }
+        if(!utilModel.isConnected())
+            return Completable.error(new RuntimeException(Constants.NO_NETWORK));
 
         return eventService
-                .login(new Login(username, password))
-                .doOnNext(loginResponse -> utilModel.saveToken(loginResponse.getAccessToken()))
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .doOnComplete(() ->
-                    databaseRepository.getAllItems(User.class)
-                    .any(user -> !user.getEmail().equals(username))
-                    .filter(differentUser -> differentUser)
-                    .subscribeOn(Schedulers.io())
-                    .subscribe(differentUser ->
-                        utilModel.deleteDatabase()
-                            .subscribe()
-                    ));
+            .login(new Login(username, password))
+            .flatMapSingle(loginResponse -> {
+                utilModel.saveToken(loginResponse.getAccessToken());
+                utilModel.addStringSetElement(Constants.SHARED_PREFS_SAVED_EMAIL, username);
+
+                return isPreviousUser();
+            })
+            .flatMapCompletable(isPrevious -> {
+                if (!isPrevious)
+                    return utilModel.deleteDatabase();
+
+                return Completable.complete();
+            })
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread());
+    }
+
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    public Single<Boolean> isPreviousUser() {
+        return databaseRepository.getAllItems(User.class)
+            .first(EMPTY)
+            .flatMap(user -> {
+                if (user.equals(EMPTY))
+                    return Single.just(false);
+
+                return eventRepository.getOrganiser(true)
+                    .firstElement()
+                    .map(user::equals)
+                    .toSingle();
+            });
     }
 
     @Override
